+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Only major refactoring improvements or functional changes should be noted.
 ## [0.20.1-dev] - 2024-10-xx - Pipeline changes
 #### Fixed
 #### Changed
+- raw files cleaner keeps normal raws
 #### New
 - Sutime DURATION analysis